+++ main/core/Resources/modules/tools/home/editor/containers/main.jsx
@@ -1,14 +1,9 @@
 import {connect} from 'react-redux'
-import merge from 'lodash/merge'
 
-import {trans} from '#/main/app/intl/translation'
-import {makeId} from '#/main/core/scaffolding/id'
 import {withRouter} from '#/main/app/router'
 import {actions as formActions} from '#/main/app/content/form/store/actions'
 import {selectors as securitySelectors} from '#/main/app/security/store'
 import {selectors as toolSelectors} from '#/main/core/tool/store'
-
-import {Tab as TabTypes} from '#/main/core/tools/home/prop-types'
 
 import {EditorMain as EditorMainComponent} from '#/main/core/tools/home/editor/components/main'
 import {actions, selectors} from '#/main/core/tools/home/store'
@@ -39,22 +34,8 @@ const EditorMain = withRouter(
       setErrors(errors) {
         dispatch(formActions.setErrors(editorSelectors.FORM_NAME, errors))
       },
-      createTab(context, administration, position, navigate){
-        const newTabId = makeId()
-
-        dispatch(formActions.updateProp(editorSelectors.FORM_NAME, `[${position}]`, merge({}, TabTypes.defaultProps, {
-          id: newTabId,
-          title: trans('tab'),
-          longTitle: trans('tab'),
-          position: position + 1,
-          type: administration ? 'administration' : context.type,
-          administration: administration,
-          user: context.type === 'desktop' && !administration ? ownProps.currentUser : null,
-          workspace: context.type === 'workspace' ? {uuid: context.data.uuid} : null
-        })))
-
-        // open new tab
-        navigate(`${ownProps.path}/edit/tab/${newTabId}`)
+      createTab(context, administration, position, navigate) {
+        dispatch(editorActions.createTab(context, administration, position, ownProps.currentUser, (path) => navigate(ownProps.path+path)))
       },
       moveTab(tabs, currentTab, newPosition) {
         dispatch(editorActions.moveTab(tabs, currentTab, newPosition))
